+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/plan02/day01/p001/Solution977_02.java
@@ -1,4 +1,4 @@
-package com.noob.algorithm.daily.plan02.day01;
+package com.noob.algorithm.daily.plan02.day01.p001;
 
 /**
  * 🟢 977 有序数组的平方 - https://leetcode.cn/problems/squares-of-a-sorted-array/description/
