+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/plan02/day01/p002/Solution142_02.java
@@ -1,5 +1,5 @@
 
-package com.noob.algorithm.daily.plan02.day01;
+package com.noob.algorithm.daily.plan02.day01.p002;
 
 import com.noob.algorithm.daily.base.ListNode;
