+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/plan02/day01/p002/Solution160_02.java
@@ -1,4 +1,4 @@
-package com.noob.algorithm.daily.plan02.day01;
+package com.noob.algorithm.daily.plan02.day01.p002;
 
 import com.noob.algorithm.daily.base.ListNode;
